Added Grid feature

+++ src/Utils/Box.java
@@ -5,10 +5,10 @@ import Utils.Pixel;
 public class Box {
 
     public Pixel pos;
-    public int width;
-    public int height;
+    public double width;
+    public double height;
 
-    public Box(Pixel pos, int width, int height) {
+    public Box(Pixel pos, double width, double height) {
         this.pos = pos;
         this.width = width;
         this.height = height;
